Reorganize DTOs and improve code formatting
Normalize code formatting, organize request and response DTOs into separate folders, and optimize imports.
