+++ src/main/java/ar/com/mercadolibre/socialmeli/dto/response/CreatePromoResponseDTO.java
@@ -1,4 +1,4 @@
-package ar.com.mercadolibre.socialmeli.dto;
+package ar.com.mercadolibre.socialmeli.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
