+++ src/main/java/ar/com/mercadolibre/socialmeli/service/IProductService.java
@@ -1,9 +1,9 @@
 package ar.com.mercadolibre.socialmeli.service;
 
-import ar.com.mercadolibre.socialmeli.dto.CreatePromoRequestDTO;
-import ar.com.mercadolibre.socialmeli.dto.CreatePromoResponseDTO;
+import ar.com.mercadolibre.socialmeli.dto.request.CreatePromoRequestDTO;
 import ar.com.mercadolibre.socialmeli.dto.request.PostDTO;
 import ar.com.mercadolibre.socialmeli.dto.request.PostsFollowersListDTO;
+import ar.com.mercadolibre.socialmeli.dto.response.CreatePromoResponseDTO;
 import ar.com.mercadolibre.socialmeli.dto.response.PostOkDTO;
 import ar.com.mercadolibre.socialmeli.dto.response.ProductPromoCountDTO;
