+++ src/main/java/ar/com/mercadolibre/socialmeli/service/IUserService.java
@@ -1,9 +1,10 @@
 package ar.com.mercadolibre.socialmeli.service;
 
-import ar.com.mercadolibre.socialmeli.dto.UserFollowedDTO;
+import ar.com.mercadolibre.socialmeli.dto.response.UserFollowedDTO;
 import ar.com.mercadolibre.socialmeli.dto.response.UserFollowerCountDTO;
 import ar.com.mercadolibre.socialmeli.dto.response.UserFollowerListDTO;
 import ar.com.mercadolibre.socialmeli.dto.response.UserOkDTO;
+
 import java.util.List;
 
 public interface IUserService {
