+++ src/main/java/ar/com/mercadolibre/socialmeli/service/impl/UserServiceImpl.java
@@ -1,27 +1,17 @@
 package ar.com.mercadolibre.socialmeli.service.impl;
 
-import ar.com.mercadolibre.socialmeli.dto.UserFollowedDTO;
-import ar.com.mercadolibre.socialmeli.dto.UserFollowedListDTO;
-import ar.com.mercadolibre.socialmeli.dto.request.PostDTO;
-import ar.com.mercadolibre.socialmeli.dto.request.PostsFollowersListDTO;
-import ar.com.mercadolibre.socialmeli.dto.response.UserFollowerCountDTO;
-import ar.com.mercadolibre.socialmeli.dto.response.UserFollowerListDTO;
-import ar.com.mercadolibre.socialmeli.dto.response.UserNameDTO;
-import ar.com.mercadolibre.socialmeli.dto.response.UserOkDTO;
-import ar.com.mercadolibre.socialmeli.entity.Post;
+import ar.com.mercadolibre.socialmeli.dto.response.*;
 import ar.com.mercadolibre.socialmeli.entity.User;
 import ar.com.mercadolibre.socialmeli.exception.BadRequestException;
 import ar.com.mercadolibre.socialmeli.exception.NotFoundException;
 import ar.com.mercadolibre.socialmeli.repository.IRepository;
 import ar.com.mercadolibre.socialmeli.service.IUserService;
-import ar.com.mercadolibre.socialmeli.utils.Utils;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import org.springframework.stereotype.Service;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -34,10 +24,13 @@ public class UserServiceImpl implements IUserService {
 
     @Override
     public List<UserFollowedDTO> findByFollowed(Integer userId, String order) {
+
         User user = repository.getUserById(userId);
+
         if (user == null) {
             throw new NotFoundException("User ID: " + userId + " doesn't exist.");
         }
+
         List<Integer> follows = user.getFollowedIds();
         List<User> allUsers = repository.getUsers();
 
@@ -68,9 +61,11 @@ public class UserServiceImpl implements IUserService {
 
         List<UserFollowedDTO> userFollowedDTOList = new ArrayList<>();
         userFollowedDTOList.add(new UserFollowedDTO(user.getUserId(), user.getUserName(), followedList));
+
         return userFollowedDTOList;
     }
 
+    @Override
     public UserFollowerListDTO getFollowerList(Integer userId, String order){
 
         if (userId == null || userId <= 0){
@@ -97,13 +92,13 @@ public class UserServiceImpl implements IUserService {
                 throw new BadRequestException("Order " + order + " not recognized.");
             }
         }
+
         User user = repository.getUserById(userId);
 
         return new UserFollowerListDTO(userId, user.getUserName(), followers);
-
     }
 
-
+    @Override
     public UserFollowerCountDTO getFollowerCount(Integer userId){
 
         if (userId == null || userId <= 0){
@@ -126,11 +121,16 @@ public class UserServiceImpl implements IUserService {
 
     @Override
     public UserOkDTO followASpecificUserById(Integer userId, Integer userIdToFollow) {
+
         if (userIdToFollow == null ||userId == null || userIdToFollow < 0 || userId < 0 || userId.equals(userIdToFollow)) {
             throw new BadRequestException("Invalid IDs");
         }
-        User user = repository.getUserById(userId);
 
+        if (!repository.existId(userIdToFollow)){
+            throw new BadRequestException("User to follow ID: " + userId + " doesn't exist.");
+        }
+
+        User user = repository.getUserById(userId);
         User userToFollow = repository.getUserById(userIdToFollow);
 
         if (userToFollow.getPosts() == null || userToFollow.getPosts().isEmpty()){
@@ -145,18 +145,15 @@ public class UserServiceImpl implements IUserService {
             throw new BadRequestException("User ID: " + userId + " already follows User ID: " + userIdToFollow);
         }
 
-
-
         user.addFollowedId(userIdToFollow);
-
         repository.updateUser(user);
 
         return new UserOkDTO("OK");
-
     }
 
     @Override
     public UserOkDTO unfollowASpecificUserById(Integer userId, Integer userIdToUnfollow) {
+
         if (userId == null || userId <= 0 || !repository.existId(userId) ){
             throw new BadRequestException("Invalid User ID: " +userId);
         }
@@ -176,7 +173,6 @@ public class UserServiceImpl implements IUserService {
         }
 
         user.removeFollowedId(userIdToUnfollow);
-
         repository.updateUser(user);
 
         return new UserOkDTO("OK");
