+++ src/main/java/ar/com/mercadolibre/socialmeli/utils/Utils.java
@@ -1,7 +1,7 @@
 package ar.com.mercadolibre.socialmeli.utils;
 
-import ar.com.mercadolibre.socialmeli.dto.request.PostDTO;
 import ar.com.mercadolibre.socialmeli.dto.PostFollowDTO;
+import ar.com.mercadolibre.socialmeli.dto.request.PostDTO;
 import ar.com.mercadolibre.socialmeli.entity.Post;
 import ar.com.mercadolibre.socialmeli.entity.Product;
 import ar.com.mercadolibre.socialmeli.entity.User;
